pyproject.toml
[tool.poetry]
name = "returnsuite-website"
version = "0.1.0"
description = "Landing website for returnsuite.com."
authors = ["Development <development@returnsuite.com>"]
license = "Proprietary"
packages = [{include = "returnsuite_website", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
Jinja2 = "^3.0.3"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
pydantic = {extras = ["email", "dotenv"], version = "^2.5.2"}
pydantic-settings = "^2.1.0"
loguru = "^0.7.2"
python-multipart = "^0.0.9"
Babel = "^2.9.1"
SQLAlchemy = "^2.0.12"
mysql-connector-python = "^8.1.0"
python-i18n = "^0.3.9"
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
aiosmtpd = "^1.4.5"
black = "^22.10.0"
djlint = "^1.34.1"
flake8 = "^6.0.0"
httpx = "^0.23.3"
isort = "^5.10.1"
livereload = "^2.6.3"
pytest = "^7.2.0"
pytest-env = "^0.8.1"

[tool.poetry.scripts]
server = "returnsuite_website.main:development_server"

[tool.black]
include = '\.pyi?$'
exclude = '''
/(
  \.git
  | backup
  | dist
)/
'''

[tool.djlint]
extension="html.jinja2"
format_attribute_template_tags=true
format_js=true
indent=2
max_attribute_length=320
max_blank_lines=1
max_line_length=120
profile="jinja"

[tool.djlint.js]
indent_size=2

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
env = ["APP_ENV=test"]
pythonpath = ["src"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
